#125 - Making unittests independent

+++ source/tools/configuration.py
@@ -20,6 +20,20 @@ import etcd
 import json
 import random
 import string
+import unittest
+from itertools import groupby
+try:
+    from requests.packages.urllib3 import disable_warnings
+except ImportError:
+    import requests
+    try:
+        reload(requests)  # Required for 2.6 > 2.7 upgrade (new requests.packages module)
+    except ImportError:
+        pass  # So, this reload fails because of some FileNodeWarning that can't be found. But it did reload. Yay.
+    from requests.packages.urllib3 import disable_warnings
+from requests.packages.urllib3.exceptions import InsecurePlatformWarning
+from requests.packages.urllib3.exceptions import InsecureRequestWarning
+from requests.packages.urllib3.exceptions import SNIMissingWarning
 
 
 class EtcdConfiguration(object):
@@ -45,6 +59,11 @@ class EtcdConfiguration(object):
         > print EtcdConfiguration.get('/bar')
         < {u'a': {u'b': u'test'}}
     """
+    _unittest_data = {}
+
+    disable_warnings(InsecurePlatformWarning)
+    disable_warnings(InsecureRequestWarning)
+    disable_warnings(SNIMissingWarning)
 
     def __init__(self):
         """
@@ -149,24 +168,6 @@ class EtcdConfiguration(object):
         return EtcdConfiguration._dir_exists(key)
 
     @staticmethod
-    def create_dir(key):
-        """
-        Creates a directory, including subdirs
-        :param key: full directory path to be created
-        :return: None
-        """
-        EtcdConfiguration._create_dir(key)
-
-    @staticmethod
-    def delete_dir(key):
-        """
-        Deletes a directory
-        :param key: full directory path to be deleted
-        :return: None
-        """
-        EtcdConfiguration._delete_dir(key)
-
-    @staticmethod
     def list(key):
         """
         List all keys in tree
@@ -210,6 +211,19 @@ class EtcdConfiguration(object):
 
     @staticmethod
     def _dir_exists(key):
+        key = EtcdConfiguration._coalesce_dashes(key=key)
+
+        # Unittests
+        if hasattr(unittest, 'running_tests') and getattr(unittest, 'running_tests') is True:
+            stripped_key = key.strip('/')
+            current_dict = EtcdConfiguration._unittest_data
+            for part in stripped_key.split('/'):
+                if part not in current_dict or not isinstance(current_dict[part], dict):
+                    return False
+                current_dict = current_dict[part]
+            return True
+
+        # Real implementation
         try:
             client = EtcdConfiguration._get_client()
             return client.get(key).dir
@@ -217,19 +231,30 @@ class EtcdConfiguration(object):
             return False
 
     @staticmethod
-    def _create_dir(key):
-        if not EtcdConfiguration._dir_exists(key):
-            client = EtcdConfiguration._get_client()
-            client.write(key, '', dir=True)
-
-    @staticmethod
-    def _delete_dir(key):
-        if EtcdConfiguration._dir_exists(key):
-            client = EtcdConfiguration._get_client()
-            client.delete(key, dir=True)
-
-    @staticmethod
     def _list(key):
+        key = EtcdConfiguration._coalesce_dashes(key=key)
+
+        # Unittests
+        if hasattr(unittest, 'running_tests') and getattr(unittest, 'running_tests') is True:
+            data = EtcdConfiguration._unittest_data
+            ends_with_dash = key.endswith('/')
+            starts_with_dash = key.startswith('/')
+            stripped_key = key.strip('/')
+            for part in stripped_key.split('/'):
+                if part not in data:
+                    raise etcd.EtcdKeyNotFound('Key not found: {0}'.format(key))
+                data = data[part]
+            if data:
+                for sub_key in data:
+                    if ends_with_dash is True:
+                        yield '/{0}/{1}'.format(stripped_key, sub_key)
+                    else:
+                        yield sub_key if starts_with_dash is True else '/{0}'.format(sub_key)
+            elif starts_with_dash is False or ends_with_dash is True:
+                yield '/{0}'.format(stripped_key)
+            return
+
+        # Real implementation
         client = EtcdConfiguration._get_client()
         for child in client.get(key).children:
             if child.key is not None and child.key != key:
@@ -237,25 +262,89 @@ class EtcdConfiguration(object):
 
     @staticmethod
     def _delete(key, recursive):
+        key = EtcdConfiguration._coalesce_dashes(key=key)
+
+        # Unittests
+        if hasattr(unittest, 'running_tests') and getattr(unittest, 'running_tests') is True:
+            stripped_key = key.strip('/')
+            data = EtcdConfiguration._unittest_data
+            for part in stripped_key.split('/')[:-1]:
+                if part not in data:
+                    raise etcd.EtcdKeyNotFound('Key not found : {0}'.format(key))
+                data = data[part]
+            key_to_remove = stripped_key.split('/')[-1]
+            if key_to_remove in data:
+                del data[key_to_remove]
+            return
+
+        # Real implementation
         client = EtcdConfiguration._get_client()
         client.delete(key, recursive=recursive)
 
     @staticmethod
     def _get(key, raw):
-        client = EtcdConfiguration._get_client()
-        data = client.read(key).value
+        key = EtcdConfiguration._coalesce_dashes(key=key)
+
+        # Unittests
+        if hasattr(unittest, 'running_tests') and getattr(unittest, 'running_tests') is True:
+            if key in ['', '/']:
+                return
+            stripped_key = key.strip('/')
+            data = EtcdConfiguration._unittest_data
+            for part in stripped_key.split('/')[:-1]:
+                if part not in data:
+                    raise etcd.EtcdKeyNotFound('Key not found : {0}'.format(key))
+                data = data[part]
+            last_part = stripped_key.split('/')[-1]
+            if last_part not in data:
+                raise etcd.EtcdKeyNotFound('Key not found : {0}'.format(key))
+            data = data[last_part]
+            if isinstance(data, dict):
+                data = None
+        else:
+            # Real implementation
+            client = EtcdConfiguration._get_client()
+            data = client.read(key).value
+
         if raw is True:
             return data
         return json.loads(data)
 
     @staticmethod
     def _set(key, value, raw):
-        client = EtcdConfiguration._get_client()
+        key = EtcdConfiguration._coalesce_dashes(key=key)
         data = value
         if raw is False:
             data = json.dumps(value)
+
+        # Unittests
+        if hasattr(unittest, 'running_tests') and getattr(unittest, 'running_tests') is True:
+            stripped_key = key.strip('/')
+            ut_data = EtcdConfiguration._unittest_data
+            for part in stripped_key.split('/')[:-1]:
+                if part not in ut_data:
+                    ut_data[part] = {}
+                ut_data = ut_data[part]
+
+            ut_data[stripped_key.split('/')[-1]] = data
+            return
+
+        # Real implementation
+        client = EtcdConfiguration._get_client()
         client.write(key, data)
 
     @staticmethod
     def _get_client():
         return etcd.Client(port=2379, use_proxies=True)
+
+    @staticmethod
+    def _coalesce_dashes(key):
+        """
+        Remove multiple dashes, eg: //ovs//framework/ becomes /ovs/framework/
+        :param key: Key to convert
+        :type key: str
+
+        :return: Key without multiple dashes after one another
+        :rtype: str
+        """
+        return ''.join(k if k == '/' else ''.join(group) for k, group in groupby(key))
